projects/JGroupsTester/results/requirements.txt
bitmath==1.3.1.2
matplotlib==1.5.3
numpy==1.11.2
pandas==0.19.1
pandas-datareader==0.2.1
progressbar2==3.10.1
tqdm==4.8.4